Backend/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flatbuffers==23.3.3
gast==0.4.0
google-api-core==2.11.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
joblib==1.2.0
keras==2.11.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pickle5==0.0.11
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
python-decouple==3.8
pytz==2022.7.1
redis==4.5.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
termcolor==2.2.0
threadpoolctl==3.1.0
twilio==7.17.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
xgboost==1.7.4
zipp==3.15.0
